Final tests

+++ src/main/java/wobbly/pigeons/expensemanager/util/ConverterRestClient.java
@@ -23,10 +23,9 @@ public class ConverterRestClient {
     }
 
 
-    public Long getConversionAmount (String baseCode, String targetCode, double amount){
+    public Double getConversionAmount (String baseCode, String targetCode, double amount){
 
-     //   return (Double) webClient
-        return (Long) webClient
+        return (Double) webClient
                 .get()
                 // .uri("/pair/{base_code}/target_code/{amount}")
                 .uri (uriBuilder -> uriBuilder
